requirements.txt
airtable-python-wrapper==0.15.3
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.3.2
click==8.2.1
cloudscraper==1.2.71
colorama==0.4.6
filelock==3.13.1
fsspec==2024.2.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
gspread==6.2.1
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.8
Jinja2==3.1.3
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
llvmlite==0.43.0
lxml==4.9.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.4.0
mpmath==1.3.0
narwhals==1.48.1
networkx==3.2.1
notion-client==2.4.0
numba==0.60.0
numpy==1.26.3
oauth2client==4.1.3
oauthlib==3.3.1
openai-whisper==20231117
outcome==1.3.0.post0
packaging==24.2
pandas==2.3.1
pillow==10.2.0
plotly==6.2.0
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.26.0
rsa==4.9.1
selenium==4.34.2
setuptools==70.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soccerdata==1.8.7
sortedcontainers==2.4.0
soupsieve==2.7
streamlit==1.47.0
sympy==1.12
tenacity==9.1.2
tiktoken==0.7.0
toml==0.10.2
torch==2.4.0+cu124
torchaudio==2.4.0+cu124
torchvision==0.19.0+cu124
tornado==6.5.1
tqdm==4.66.5
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
undetected-chromedriver==3.5.5
unicode==2.9
Unidecode==1.4.0
uritemplate==4.2.0
urllib3==2.5.0
watchdog==6.0.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
youtube-transcript-api==0.6.2
